add a reset indices command and test to verify it

+++ gateway/sds_gateway/api_methods/utils/metadata_schemas.py
@@ -221,7 +221,7 @@ rh_capture_index_mapping = {
         "type": "float",
     },
     "altitude": {
-        "type": "double",
+        "type": "float",
     },
     "mac_address": {
         "type": "keyword",
